Move contents of config/set to configOptions.ts
